Genie.Commands.called_command — Function

called_command(args::Dict, key::String) :: Bool

Checks whether or not a certain command was invoked by looking at the command line args.

Genie.Commands.execute — Function

execute(config::Settings) :: Nothing

Runs the requested Genie app command, based on the args passed to the script.

Genie.Commands.parse_commandline_args — Function

parse_commandline_args() :: Dict{String,Any}

Extracts the command line args passed into the app and returns them as a Dict, possibly setting up defaults. Also, it is used by the ArgParse module to populate the command line help for the app -h.