Genie.Configuration.buildpath — Function

buildpath() :: String

Constructs the temp dir where Genie's view files are built.

Genie.Configuration.config! — Function

config!(; kwargs...)

Updates Genie.config using the provided keyword arguments.

Genie.Configuration.env — Function

env() :: String

Returns the current Genie environment.

Examples

julia> Configuration.env()
"dev"

Genie.Configuration.isdev — Function

isdev()  :: Bool

Set of utility functions that return whether or not the current environment is development, production or testing.

Examples

julia> Configuration.isdev()
true

julia> Configuration.isprod()
false

Genie.Configuration.isprod — Function

isprod() :: Bool

Set of utility functions that return whether or not the current environment is development, production or testing.

Examples

julia> Configuration.isdev()
true

julia> Configuration.isprod()
false

Genie.Configuration.istest — Function

istest() :: Bool

Set of utility functions that return whether or not the current environment is development, production or testing.

Examples

julia> Configuration.isdev()
true

julia> Configuration.isprod()
false

Genie.Configuration.pkginfo — Function

pkginfo(pkg::String)

Returns installed package information for pkg

Genie.Configuration.Settings — Type

mutable struct Settings

App configuration - sets up the app's defaults. Individual options are overwritten in the corresponding environment file.

Arguments

server_port::Int: the port for running the web server (default 8000)

server_host::String: the host for running the web server (default "127.0.0.1")

server_document_root::String: path to the document root (default "public/")

server_handle_static_files::Bool: if true, Genie will also serve static files. In production, it is recommended to serve static files with a web server like Nginx.

server_signature::String: Genie's signature used for tagging the HTTP responses. If empty, it will not be added.

app_env::String: the environment in which the app is running (dev, test, or prod)

cors_headers::Dict{String,String}: default Access-Control-* CORS settings

cors_allowed_origins::Vector{String}: allowed origin hosts for CORS settings

log_level::Logging.LogLevel: logging severity level

log_to_file::Bool: if true, information will be logged to file besides REPL

log_requests::Bool: if true, requests will be automatically logged

inflector_irregulars::Vector{Tuple{String,String}}: additional irregular singular-plural forms to be used by the Inflector

run_as_server::Bool: when true the server thread is launched synchronously to avoid that the script exits

websockets_server::Bool: if true, the websocket server is also started together with the web server

websockets_port::Int: the port for the websocket server (default server_port)

initializers_folder::String: the folder where the initializers are located (default "initializers/")

path_config::String: the path to the configurations folder (default "config/")

path_env::String: the path to the environment files (default "<path_config>/env/")

path_app::String: the path to the app files (default "app/")

html_parser_close_tag::String: default " /". Can be changed to an empty string "" so the single tags would not be closed.

webchannels_keepalive_frequency::Int: default 30000. Frequency in milliseconds to send keepalive messages to webchannel/websocket to keep the connection alive. Set to 0 to disable keepalive messages.