Genie.Loader.autoload — Function

autoload

Automatically and recursively includes files from the indicated root_dir into the indicated context module, skipping directories from dir. The files are set up with Revise to be automatically reloaded when changed (in dev environment).

Genie.Loader.bootstrap — Function

bootstrap(context::Union{Module,Nothing} = nothing) :: Nothing

Kickstarts the loading of a Genie app by loading the environment settings.

Genie.Loader.default_context — Function

default_context(context::Union{Module,Nothing})

Sets the module in which the code is loaded (the app's module)

Missing docstring.

Missing docstring for importenv. Check Documenter's build log for details.

Genie.Loader.load — Function

load(; context::Union{Module,Nothing} = nothing) :: Nothing

Main entry point to loading a Genie app.

Genie.Loader.load_helpers — Function

load_helpers(root_dir::String = Genie.config.path_helpers) :: Nothing

Automatically recursively includes files from helpers/ and subfolders.

Genie.Loader.load_initializers — Function

load_initializers(root_dir::String = Genie.config.path_config; context::Union{Module,Nothing} = nothing) :: Nothing

Automatically recursively includes files from initializers/ and subfolders.

Genie.Loader.load_libs — Function

load_libs(root_dir::String = Genie.config.path_lib) :: Nothing

Recursively includes files from lib/ and subfolders. The lib/ folder, if present, is designed to host user code in the form of .jl files.

Genie.Loader.load_plugins — Function

load_plugins(root_dir::String = Genie.config.path_plugins; context::Union{Module,Nothing} = nothing) :: Nothing

Automatically recursively includes files from plugins/ and subfolders.

Genie.Loader.load_resources — Function

load_resources(root_dir::String = Genie.config.path_resources) :: Nothing

Automatically recursively includes files from resources/ and subfolders.

Genie.Loader.load_routes — Function

load_routes(routes_file::String = Genie.ROUTES_FILE_NAME; context::Union{Module,Nothing} = nothing) :: Nothing

Loads the routes file.