Missing docstring.

Missing docstring for TaskInfo. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TaskResult. Check Documenter's build log for details.

Missing docstring.

Missing docstring for tasks. Check Documenter's build log for details.

Missing docstring.

Missing docstring for VoidTaskResult. Check Documenter's build log for details.

Genie.Toolbox.validtaskname — Function

validtaskname(task_name::String) :: String

Attempts to convert a potentially invalid (partial) task_name into a valid one.

Genie.Toolbox.taskdocs — Function

task_docs(module_name::Module) :: String

Retrieves the docstring of the runtask method and returns it as a string.

Genie.Toolbox.loadtasks — Function

loadtasks(; filter_type_name = Symbol()) :: Vector{TaskInfo}

Returns a vector of all registered Genie tasks.

Genie.Toolbox.printtasks — Function

Prints a list of all the registered Genie tasks to the standard output.

new — Keyword

new, or new{A,B,...}

Special function available to inner constructors which creates a new object of the type. The form new{A,B,...} explicitly specifies values of parameters for parametric types. See the manual section on Inner Constructor Methods for more information.

Genie.Toolbox.taskfilename — Function

task_file_name(cmd_args::Dict{String,Any}, config::Settings) :: String

Computes the name of a Genie task based on the command line input.

Genie.Toolbox.taskmodulename — Function

task_module_name(underscored_task_name::String) :: String

Computes the name of a Genie task based on the command line input.

Genie.Toolbox.isvalidtask! — Function

isvalidtask!(parsed_args::Dict{String,Any}) :: Dict{String,Any}

Checks if the name of the task passed as the command line arg is valid task identifier – if not, attempts to address it, by appending the TASKSUFFIX suffix. Returns the potentially modified `parsedargsDict`.