Missing docstring.

Missing docstring for CLIENTS. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MESSAGE_QUEUE. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SUBSCRIPTIONS. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ChannelClient. Check Documenter's build log for details.

Genie.WebThreads.ChannelClientsCollection — Type

Dict([itr])

Dict{K,V}() constructs a hash table with keys of type K and values of type V. Keys are compared with isequal and hashed with hash.

Given a single iterable argument, constructs a Dict whose key-value pairs are taken from 2-tuples (key,value) generated by the argument.

Examples

julia> Dict([("A", 1), ("B", 2)])
Dict{String, Int64} with 2 entries:
  "B" => 2
  "A" => 1

Alternatively, a sequence of pair arguments may be passed.

julia> Dict("A"=>1, "B"=>2)
Dict{String, Int64} with 2 entries:
  "B" => 2
  "A" => 1

Missing docstring.

Missing docstring for ChannelMessage. Check Documenter's build log for details.

Genie.WebThreads.ChannelSubscriptionsCollection — Type

Dict([itr])

Dict{K,V}() constructs a hash table with keys of type K and values of type V. Keys are compared with isequal and hashed with hash.

Given a single iterable argument, constructs a Dict whose key-value pairs are taken from 2-tuples (key,value) generated by the argument.

Examples

julia> Dict([("A", 1), ("B", 2)])
Dict{String, Int64} with 2 entries:
  "B" => 2
  "A" => 1

Alternatively, a sequence of pair arguments may be passed.

julia> Dict("A"=>1, "B"=>2)
Dict{String, Int64} with 2 entries:
  "B" => 2
  "A" => 1

Genie.WebThreads.ClientId — Type

UInt64 <: Unsigned

64-bit unsigned integer type.

Genie.WebThreads.ChannelName — Type

String(v::AbstractVector{UInt8})

Create a new String object from a byte vector v containing UTF-8 encoded characters. If v is Vector{UInt8} it will be truncated to zero length and future modification of v cannot affect the contents of the resulting string. To avoid truncation of Vector{UInt8} data, use String(copy(v)); for other AbstractVector types, String(v) already makes a copy.

When possible, the memory of v will be used without copying when the String object is created. This is guaranteed to be the case for byte vectors returned by take! on a writable IOBuffer and by calls to read(io, nb). This allows zero-copy conversion of I/O data to strings. In other cases, Vector{UInt8} data may be copied, but v is truncated anyway to guarantee consistent behavior.

String(s::AbstractString)

Convert a string to a contiguous byte array representation encoded as UTF-8 bytes. This representation is often appropriate for passing strings to C.

Missing docstring.

Missing docstring for MessagePayload. Check Documenter's build log for details.

Genie.WebThreads.broadcast — Function

Pushes msg (and payload) to all the clients subscribed to the channels in channels.

Pushes msg (and payload) to all the clients subscribed to all the channels.

Missing docstring.

Missing docstring for channels. Check Documenter's build log for details.

Missing docstring.

Missing docstring for clients. Check Documenter's build log for details.

Missing docstring.

Missing docstring for connected_clients. Check Documenter's build log for details.

Missing docstring.

Missing docstring for disconnected_clients. Check Documenter's build log for details.

Genie.WebThreads.message — Function

Pushes msg (and payload) to channel.

Writes msg to message queue for client.

Genie.WebThreads.pop_subscription — Function

Removes the subscription of client to channel.

Removes all subscriptions of client.

Missing docstring.

Missing docstring for pull. Check Documenter's build log for details.

Missing docstring.

Missing docstring for push. Check Documenter's build log for details.

Genie.WebThreads.push_subscription — Function

Adds a new subscription for client to channel.

Genie.WebThreads.subscribe — Function

Subscribes a web thread client wt to channel.

Missing docstring.

Missing docstring for subscriptions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for timestamp_client. Check Documenter's build log for details.

Genie.WebThreads.unsubscribe — Function

Unsubscribes a web socket client wt from channel.

Genie.WebThreads.unsubscribe_client — Function

Unsubscribes a web socket client wt from all the channels.

Missing docstring.

Missing docstring for unsubscribe_clients. Check Documenter's build log for details.

Genie.WebThreads.unsubscribe_disconnected_clients — Function

unsubscribedisconnectedclients() :: ChannelClientsCollection

Unsubscribes clients which are no longer connected.

Missing docstring.

Missing docstring for webthreads. Check Documenter's build log for details.