-

Genie

The highly productive Julia web framework

Documentation

Acknowledgements

Genie.Cache.withcache — Function.

withcache(f::Function, key::Union{String,Symbol}, expiration::Int = CACHE_DURATION; dir = "", condition::Bool = true)

Executes the function f and stores the result into the cache for the duration (in seconds) of expiration. Next time the function is invoked, if the cache has not expired, the cached result is returned skipping the function execution. The optional dir param is used to designate the folder where the cache will be stored (within the configured cache folder). If condition is false caching will be skipped.

source

Genie.Cache.purge — Function.

purge(key::Union{String,Symbol}; dir::String = "") :: Nothing

Removes the cache data stored under the key key.

source

Genie.Cache.purgeall — Function.

purgeall(; dir::String = "") :: Nothing

Removes all cached data.

source