-

Genie

The highly productive Julia web framework

Documentation

Acknowledgements

Missing docstring.

Missing docstring for ChannelClient. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ChannelMessage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for clients. Check Documenter's build log for details.

Missing docstring.

Missing docstring for subscriptions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for websockets. Check Documenter's build log for details.

Missing docstring.

Missing docstring for channels. Check Documenter's build log for details.

Genie.WebChannels.connected_clients — Function.

source

Genie.WebChannels.disconnected_clients — Function.

source

Genie.WebChannels.subscribe — Function.

Subscribes a web socket client ws to channel.

source

Genie.WebChannels.unsubscribe — Function.

Unsubscribes a web socket client ws from channel.

source

Genie.WebChannels.unsubscribe_client — Function.

Unsubscribes a web socket client ws from all the channels.

source

Missing docstring.

Missing docstring for unsubscribe_disconnected_clients. Check Documenter's build log for details.

Genie.WebChannels.push_subscription — Function.

Adds a new subscription for client to channel.

source

Genie.WebChannels.pop_subscription — Function.

Removes the subscription of client to channel.

source

Removes all subscriptions of client.

source

Genie.WebChannels.broadcast — Function.

Pushes msg (and payload) to all the clients subscribed to the channels in channels.

source

Pushes msg (and payload) to all the clients subscribed to all the channels.

source

Genie.WebChannels.message — Function.

Pushes msg (and payload) to channel.

source

Writes msg to web socket for client.

source