-

Genie

The highly productive Julia web framework

Documentation

Acknowledgements

Genie.REPL.secret_token — Function.

secret_token() :: String

Generates a random secret token to be used for configuring the SECRET_TOKEN const.

source

Genie.REPL.copy_fullstack_app — Function.

source

Genie.REPL.copy_microstack_app — Function.

source

Genie.REPL.copy_db_support — Function.

source

Genie.REPL.copy_mvc_support — Function.

source

Genie.REPL.write_secrets_file — Function.

Generates a valid secrets.jl file with a random SECRET_TOKEN.

source

Genie.REPL.write_app_custom_files — Function.

source

Genie.REPL.install_app_dependencies — Function.

source

Genie.REPL.autostart_app — Function.

source

Genie.REPL.newapp — Function.

newapp(path::String = "."; autostart::Bool = true, fullstack::Bool = false, dbsupport::Bool = false, mvcsupport::Bool = false) :: Nothing

Scaffolds a new Genie app, setting up the file structure indicated by the various arguments.

Arguments

path::String: the name of the app and the path where to bootstrap it

autostart::Bool: automatically start the app once the file structure is created

fullstack::Bool: the type of app to be bootstrapped. The fullstack app includes MVC structure, DB connection code, and asset pipeline files.

dbsupport::Bool: bootstrap the files needed for DB connection setup via the SearchLight ORM

mvcsupport::Bool: adds the files used for Flax view templates rendering and working with resources

Examples

julia> Genie.newapp("MyGenieApp")
2019-08-06 16:54:15:INFO:Main: Done! New app created at MyGenieApp
2019-08-06 16:54:15:DEBUG:Main: Changing active directory to MyGenieApp
2019-08-06 16:54:15:DEBUG:Main: Installing app dependencies
 Resolving package versions...
  Updating `~/Dropbox/Projects/GenieTests/MyGenieApp/Project.toml`
  [c43c736e] + Genie v0.10.1
  Updating `~/Dropbox/Projects/GenieTests/MyGenieApp/Manifest.toml`

2019-08-06 16:54:27:INFO:Main: Starting your brand new Genie app - hang tight!
 _____         _
|   __|___ ___|_|___
|  |  | -_|   | | -_|
|_____|___|_|_|_|___|

┌ Info:
│ Starting Genie in >> DEV << mode
└
[ Info: Logging to file at MyGenieApp/log/dev.log
[ Info: Ready!
2019-08-06 16:54:32:DEBUG:Main: Web Server starting at http://127.0.0.1:8000
2019-08-06 16:54:32:DEBUG:Main: Web Server running at http://127.0.0.1:8000

source

Genie.REPL.loadapp — Function.

loadapp(path::String = "."; autostart::Bool = false) :: Nothing

Loads an existing Genie app from the file system, within the current Julia REPL session.

Arguments

path::String: the path to the Genie app on the file system.

autostart::Bool: automatically start the app upon loading it.

Examples

shell> tree -L 1
.
├── Manifest.toml
├── Project.toml
├── bin
├── bootstrap.jl
├── config
├── env.jl
├── genie.jl
├── log
├── public
├── routes.jl
└── src

5 directories, 6 files

julia> using Genie

julia> Genie.loadapp(".")
 _____         _
|   __|___ ___|_|___
|  |  | -_|   | | -_|
|_____|___|_|_|_|___|

┌ Info:
│ Starting Genie in >> DEV << mode
└
[ Info: Logging to file at MyGenieApp/log/dev.log

source

Genie.REPL.setup_windows_bin_files — Function.

source

Missing docstring.

Missing docstring for setup_nix_bin_files. Check Documenter's build log for details.