-

Genie

The highly productive Julia web framework

Documentation

Acknowledgements

Missing docstring.

Missing docstring for ChannelClient. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ChannelMessage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for clients. Check Documenter's build log for details.

Missing docstring.

Missing docstring for subscriptions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for websockets. Check Documenter's build log for details.

Missing docstring.

Missing docstring for channels. Check Documenter's build log for details.

Missing docstring.

Missing docstring for connected_clients. Check Documenter's build log for details.

Missing docstring.

Missing docstring for disconnected_clients. Check Documenter's build log for details.

Missing docstring.

Missing docstring for subscribe. Check Documenter's build log for details.

Missing docstring.

Missing docstring for unsubscribe. Check Documenter's build log for details.

Missing docstring.

Missing docstring for unsubscribe_client. Check Documenter's build log for details.

Missing docstring.

Missing docstring for unsubscribe_disconnected_clients. Check Documenter's build log for details.

Missing docstring.

Missing docstring for push_subscription. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pop_subscription. Check Documenter's build log for details.

Base.Broadcast.broadcast — Function.

broadcast(f, As...)

Broadcast the function f over the arrays, tuples, collections, Refs and/or scalars As.

Broadcasting applies the function f over the elements of the container arguments and the scalars themselves in As. Singleton and missing dimensions are expanded to match the extents of the other arguments by virtually repeating the value. By default, only a limited number of types are considered scalars, including Numbers, Strings, Symbols, Types, Functions and some common singletons like missing and nothing. All other arguments are iterated over or indexed into elementwise.

The resulting container type is established by the following rules:

If all the arguments are scalars or zero-dimensional arrays, it returns an unwrapped scalar.

If at least one argument is a tuple and all others are scalars or zero-dimensional arrays, it returns a tuple.

All other combinations of arguments default to returning an Array, but custom container types can define their own implementation and promotion-like rules to customize the result when they appear as arguments.

A special syntax exists for broadcasting: f.(args...) is equivalent to broadcast(f, args...), and nested f.(g.(args...)) calls are fused into a single broadcast loop.

Examples

julia> A = [1, 2, 3, 4, 5]
5-element Array{Int64,1}:
 1
 2
 3
 4
 5

julia> B = [1 2; 3 4; 5 6; 7 8; 9 10]
5×2 Array{Int64,2}:
 1   2
 3   4
 5   6
 7   8
 9  10

julia> broadcast(+, A, B)
5×2 Array{Int64,2}:
  2   3
  5   6
  8   9
 11  12
 14  15

julia> parse.(Int, ["1", "2"])
2-element Array{Int64,1}:
 1
 2

julia> abs.((1, -2))
(1, 2)

julia> broadcast(+, 1.0, (0, -2.0))
(1.0, -1.0)

julia> (+).([[0,2], [1,3]], Ref{Vector{Int}}([1,-1]))
2-element Array{Array{Int64,1},1}:
 [1, 1]
 [2, 2]

julia> string.(("one","two","three","four"), ": ", 1:4)
4-element Array{String,1}:
 "one: 1"
 "two: 2"
 "three: 3"
 "four: 4"

source

Missing docstring.

Missing docstring for message. Check Documenter's build log for details.