-

Genie

The highly productive Julia web framework

Documentation

Acknowledgements

Genie.Helpers.request — Function.

request() :: HTTP.Request

Returns the Request object associated with the current HTTP request.

source

request(params::Dict{Symbol,Any}) :: HTTP.Request

Returns the Request object associated with the current HTTP request.

source

Genie.Helpers.response — Function.

response() :: HTTP.Response

Returns the Response object associated with the current HTTP request.

source

response(params::Dict{Symbol,Any}) :: HTTP.Response

Returns the Response object associated with the current HTTP request.

source

Genie.Helpers.flash — Function.

flash()

Returns the flash dict object associated with the current HTTP request.

source

flash(params::Dict{Symbol,Any})

Returns the flash dict object associated with the current HTTP request.

source

flash(value::Any) :: Nothing

Stores value on the flash.

source

flash(value::Any, params::Dict{Symbol,Any}) :: Nothing

Stores value on the flash.

source

Genie.Helpers.flash_has_message — Function.

flash_has_message() :: Bool

Checks if there's any value on the flash storage

source

Genie.Helpers.wsclient — Function.

wsclient(params::Dict{Symbol,Any}) :: HTTP.WebSockets.WebSocket

Returns the WebSocket object associated with the current WS request.

source