Genie

The highly productive Julia web framework

Documentation

Acknowledgements

Genie.Plugins.recursive_copy — Function

recursive_copy(path::String, dest::String; only_hidden = true, force = false)

Utility function to copy plugin files from package dir to app.

Genie.Plugins.congrats — Function

congrats()

Shows success message and instructions when scaffolding a plugin.

Genie.Plugins.scaffold — Function

scaffold(plugin_name::String, dest::String = "."; force = false)

Scaffolds a new plugin as a Julia project

Genie.Plugins.install — Function

install(path::String, dest::String; force = false)

Utility to allow users to install a plugin