Genie

The highly productive Julia web framework

Documentation

Acknowledgements

Missing docstring.

Missing docstring for TaskInfo. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TaskResult. Check Documenter's build log for details.

Genie.Toolbox.tasks — Function

tasks(; filter_type_name = Symbol()) :: Vector{TaskInfo}

Returns a vector of all registered Genie tasks.

Genie.Toolbox.loadtasks — Function

tasks(; filter_type_name = Symbol()) :: Vector{TaskInfo}

Returns a vector of all registered Genie tasks.

Genie.Toolbox.printtasks — Function

Prints a list of all the registered Genie tasks to the standard output.

new — Keyword

new

Special function available to inner constructors which created a new object of the type. See the manual section on Inner Constructor Methods for more information.