Genie

生産性の高いJulia Webフレームワーク

ドキュメント

謝辞

Genie.Cache.withcache — Function

withcache(f::Function, key::Union{String,Symbol}, expiration::Int = Genie.config.cache_duration; dir = "", condition::Bool = true)

Executes the function f and stores the result into the cache for the duration (in seconds) of expiration. Next time the function is invoked, if the cache has not expired, the cached result is returned skipping the function execution. The optional dir param is used to designate the folder where the cache will be stored (within the configured cache folder). If condition is false caching will be skipped.

Genie.Cache.purge — Function

purge()

Removes the cache data stored under the key key.

Genie.Cache.purgeall — Function

purgeall()

Removes all cached data.