Genie

生産性の高いJulia Webフレームワーク

ドキュメント

謝辞

Missing docstring.

Missing docstring for ChannelClient. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ChannelMessage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for clients. Check Documenter's build log for details.

Missing docstring.

Missing docstring for subscriptions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for websockets. Check Documenter's build log for details.

Missing docstring.

Missing docstring for channels. Check Documenter's build log for details.

Genie.WebChannels.connected_clients — Function

Genie.WebChannels.disconnected_clients — Function

Genie.WebChannels.subscribe — Function

Subscribes a web socket client ws to channel.

Genie.WebChannels.unsubscribe — Function

Unsubscribes a web socket client ws from channel.

Genie.WebChannels.unsubscribe_client — Function

Unsubscribes a web socket client ws from all the channels.

Missing docstring.

Missing docstring for unsubscribe_disconnected_clients. Check Documenter's build log for details.

Genie.WebChannels.push_subscription — Function

Adds a new subscription for client to channel.

Genie.WebChannels.pop_subscription — Function

Removes the subscription of client to channel.

Removes all subscriptions of client.

Genie.WebChannels.broadcast — Function

Pushes msg (and payload) to all the clients subscribed to the channels in channels.

Pushes msg (and payload) to all the clients subscribed to all the channels.

Genie.WebChannels.message — Function

Pushes msg (and payload) to channel.

Writes msg to web socket for client.