Genie.Assets.include_asset — Function

include_asset(asset_type::Union{String,Symbol}, file_name::Union{String,Symbol};
              fingerprinted::Bool = Genie.config.assets_fingerprinted) :: String

Returns the path to an asset. asset_type can be one of :js, :css. The file_name should not include the extension. fingerprinted is a Bool indicating if a fingerprint (unique hash) should be added to the asset's filename (used in production to invalidate caches).

Genie.Assets.css_asset — Function

css_asset(file_name::String; fingerprinted::Bool = Genie.config.assets_fingerprinted) :: String

Path to a css asset. The file_name should not include the extension. fingerprinted is a Bool indicating if a fingerprint (unique hash) should be added to the asset's filename (used in production to invalidate caches).

Genie.Assets.js_asset — Function

js_asset(file_name::String; fingerprinted::Bool = Genie.config.assets_fingerprinted) :: String

Path to a js asset. file_name should not include the extension. fingerprinted is a Bool indicating if a fingerprint (unique hash) should be added to the asset's filename (used in production to invalidate caches).

Genie.Assets.js_settings — Function

js_settings() :: string

Sets up a window.Genie.Settings JavaScript object which exposes relevant Genie app settings from Genie.config

Genie.Assets.embedded — Function

embeded(path::String) :: String

Reads and outputs the file at path within Genie's root package dir

Genie.Assets.channels — Function

channels() :: String

Outputs the channels.js file included with the Genie package

Genie.Assets.channels_script — Function

channels_script() :: String

Outputs the channels JavaScript content within <script>...</script> tags, for embedding into the page.

Genie.Assets.channels_support — Function

channels_support(channel = Genie.config.webchannels_default_route) :: String

Provides full web channels support, setting up routes for loading support JS files, web sockets subscription and returning the <script> tag for including the linked JS file into the web page.

Genie.Assets.favicon_support — Function

favicon_support() :: String

Outputs the <link> tag for referencing the favicon file embedded with Genie.