Genie.Generator.newapp — Function

newapp(app_name::String; autostart::Bool = true, fullstack::Bool = false, dbsupport::Bool = false, mvcsupport::Bool = false) :: Nothing

Scaffolds a new Genie app, setting up the file structure indicated by the various arguments.

Arguments

path::String: the name of the app and the path where to bootstrap it. Spaces not allowed

autostart::Bool: automatically start the app once the file structure is created

fullstack::Bool: the type of app to be bootstrapped. The fullstack app includes MVC structure, DB connection code, and asset pipeline files.

dbsupport::Bool: bootstrap the files needed for DB connection setup via the SearchLight ORM

mvcsupport::Bool: adds the files used for HTML+Julia view templates rendering and working with resources

Examples

julia> Genie.newapp("MyGenieApp")
2019-08-06 16:54:15:INFO:Main: Done! New app created at MyGenieApp
2019-08-06 16:54:15:DEBUG:Main: Changing active directory to MyGenieApp
2019-08-06 16:54:15:DEBUG:Main: Installing app dependencies
 Resolving package versions...
  Updating `~/Dropbox/Projects/GenieTests/MyGenieApp/Project.toml`
  [c43c736e] + Genie v0.10.1
  Updating `~/Dropbox/Projects/GenieTests/MyGenieApp/Manifest.toml`

2019-08-06 16:54:27:INFO:Main: Starting your brand new Genie app - hang tight!
 _____         _
|   __|___ ___|_|___
|  |  | -_|   | | -_|
|_____|___|_|_|_|___|

┌ Info:
│ Starting Genie in >> DEV << mode
└
[ Info: Logging to file at MyGenieApp/log/dev.log
[ Info: Ready!
2019-08-06 16:54:32:DEBUG:Main: Web Server starting at http://127.0.0.1:8000
2019-08-06 16:54:32:DEBUG:Main: Web Server running at http://127.0.0.1:8000

Genie.Generator.newapp_webservice — Function

newapp_webservice(path::String = "."; autostart::Bool = true, dbsupport::Bool = false) :: Nothing

Template for scaffolding a new Genie app suitable for nimble web services.

Arguments

path::String: the name of the app and the path where to bootstrap it

autostart::Bool: automatically start the app once the file structure is created

dbsupport::Bool: bootstrap the files needed for DB connection setup via the SearchLight ORM

Genie.Generator.newapp_mvc — Function

newapp_mvc(path::String = "."; autostart::Bool = true) :: Nothing

Template for scaffolding a new Genie app suitable for MVC web applications (includes MVC structure and DB support).

Arguments

path::String: the name of the app and the path where to bootstrap it

autostart::Bool: automatically start the app once the file structure is created

Genie.Generator.newapp_fullstack — Function

newapp_fullstack(path::String = "."; autostart::Bool = true) :: Nothing

Template for scaffolding a new Genie app suitable for full stack web applications (includes MVC structure, DB support, and frontend asset pipeline).

Arguments

path::String: the name of the app and the path where to bootstrap it

autostart::Bool: automatically start the app once the file structure is created

Genie.Generator.newcontroller — Function

newcontroller(resource_name::String) :: Nothing

Generates a new Genie controller file and persists it to the resources folder.

Genie.Generator.newresource — Function

newresource(resource_name::String, config::Settings) :: Nothing

Generates all the files associated with a new resource and persists them to the resources folder.

Genie.Generator.setup_resource_path — Function

setup_resource_path(resource_name::String) :: String

Computes and creates the directories structure needed to persist a new resource.

Genie.Generator.write_resource_file — Function

write_resource_file(resource_path::String, file_name::String, resource_name::String) :: Bool

Generates all resouce files and persists them to disk.

Genie.Generator.setup_windows_bin_files — Function

setup_windows_bin_files(path::String = ".") :: Nothing

Creates the bin/server and bin/repl binaries for Windows

Genie.Generator.setup_nix_bin_files — Function

setup_nix_bin_files(path::String = ".") :: Nothing

Creates the bin/server and bin/repl binaries for *nix systems

Genie.Generator.resource_does_not_exist — Function

resource_does_not_exist(resource_path::String, file_name::String) :: Bool

Returns true if the indicated resources does not exists - false otherwise.

Genie.Generator.controller_file_name — Function

controller_file_name(resource_name::Union{String,Symbol})

Computes the controller file name based on the resource name.

Genie.Generator.secret_token — Function

secret_token() :: String

Generates a random secret token to be used for configuring the call to Genie.secret_token!.

Genie.Generator.write_secrets_file — Function

write_secrets_file(app_path=".")

Generates a valid config/secrets.jl file with a random secret token.

Genie.Generator.fullstack_app — Function

fullstack_app(app_path::String = ".") :: Nothing

Writes the files necessary to create a full stack Genie app.

Genie.Generator.microstack_app — Function

microstack_app(app_path::String = ".") :: Nothing

Writes the file necessary to create a microstack app.

Genie.Generator.mvc_support — Function

mvc_support(app_path::String = ".") :: Nothing

Writes the files used for rendering resources using the MVC stack and the Genie templating system.

Genie.Generator.db_support — Function

db_support(app_path::String = ".") :: Nothing

Writes files used for interacting with the SearchLight ORM.

Genie.Generator.write_app_custom_files — Function

write_app_custom_files(path::String, app_path::String) :: Nothing

Writes the Genie app main module file.

Genie.Generator.install_app_dependencies — Function

install_app_dependencies(app_path::String = ".") :: Nothing

Installs the application's dependencies using Julia's Pkg

Genie.Generator.autostart_app — Function

autostart_app(path::String = "."; autostart::Bool = true) :: Nothing

If autostart is true, the newly generated Genie app will be automatically started.

Genie.Generator.remove_fingerprint_initializer — Function

remove_fingerprint_initializer(app_path::String = ".") :: Nothing

Removes the asset fingerprint initializers if it's not used

Genie.Generator.remove_searchlight_initializer — Function

remove_searchlight_initializer(app_path::String = ".") :: Nothing

Removes the SearchLight initializer file if it's unused