Missing docstring.

Missing docstring for ChannelNotFoundException. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ChannelClient. Check Documenter's build log for details.

Genie.WebChannels.ChannelClientsCollection — Type

Dict([itr])

Dict{K,V}() constructs a hash table with keys of type K and values of type V. Keys are compared with isequal and hashed with hash.

Given a single iterable argument, constructs a Dict whose key-value pairs are taken from 2-tuples (key,value) generated by the argument.

Examples

julia> Dict([("A", 1), ("B", 2)])
Dict{String,Int64} with 2 entries:
  "B" => 2
  "A" => 1

Alternatively, a sequence of pair arguments may be passed.

julia> Dict("A"=>1, "B"=>2)
Dict{String,Int64} with 2 entries:
  "B" => 2
  "A" => 1

Base.Dict(req::HTTP.Request) :: Dict{String,String}

Converts a HTTP.Request to a Dict.

Dict(req::Request) :: Dict{String,String}

Extracts the Cookie and Set-Cookie data from the Request and Response objects and converts it into a Dict.

Genie.WebChannels.ChannelSubscriptionsCollection — Type

Dict([itr])

Dict{K,V}() constructs a hash table with keys of type K and values of type V. Keys are compared with isequal and hashed with hash.

Given a single iterable argument, constructs a Dict whose key-value pairs are taken from 2-tuples (key,value) generated by the argument.

Examples

julia> Dict([("A", 1), ("B", 2)])
Dict{String,Int64} with 2 entries:
  "B" => 2
  "A" => 1

Alternatively, a sequence of pair arguments may be passed.

julia> Dict("A"=>1, "B"=>2)
Dict{String,Int64} with 2 entries:
  "B" => 2
  "A" => 1

Base.Dict(req::HTTP.Request) :: Dict{String,String}

Converts a HTTP.Request to a Dict.

Dict(req::Request) :: Dict{String,String}

Extracts the Cookie and Set-Cookie data from the Request and Response objects and converts it into a Dict.

Missing docstring.

Missing docstring for MessagePayload. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ChannelMessage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CLIENTS. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SUBSCRIPTIONS. Check Documenter's build log for details.

Missing docstring.

Missing docstring for clients. Check Documenter's build log for details.

Missing docstring.

Missing docstring for subscriptions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for websockets. Check Documenter's build log for details.

Missing docstring.

Missing docstring for channels. Check Documenter's build log for details.

Missing docstring.

Missing docstring for connected_clients. Check Documenter's build log for details.

Missing docstring.

Missing docstring for disconnected_clients. Check Documenter's build log for details.

Genie.WebChannels.subscribe — Function

Subscribes a web socket client ws to channel.

Genie.WebChannels.unsubscribe — Function

Unsubscribes a web socket client ws from channel.

Genie.WebChannels.unsubscribe_client — Function

Unsubscribes a web socket client ws from all the channels.

Genie.WebChannels.unsubscribe_disconnected_clients — Function

unsubscribedisconnectedclients() :: ChannelClientsCollection

Unsubscribes clients which are no longer connected.

Genie.WebChannels.push_subscription — Function

Adds a new subscription for client to channel.

Genie.WebChannels.pop_subscription — Function

Removes the subscription of client to channel.

Removes all subscriptions of client.

Genie.WebChannels.broadcast — Function

Pushes msg (and payload) to all the clients subscribed to the channels in channels.

Pushes msg (and payload) to all the clients subscribed to all the channels.

Genie.WebChannels.message — Function

Pushes msg (and payload) to channel.

Writes msg to web socket for client.