Genie

The highly productive Julia web framework

Documentation

Acknowledgements

Genie.Renderer.WebRenderable — Type

mutable struct WebRenderable

Represents an object that can be rendered on the web as a HTTP Response

Missing docstring.

Missing docstring for render. Check Documenter's build log for details.

Genie.Renderer.redirect — Function

Sets redirect headers and prepares the Response.

Genie.Renderer.hasrequested — Function

hasrequested(content_type::Symbol) :: Bool

Checks wheter or not the requested content type matches content_type.

Genie.Renderer.respond — Function

Constructs a Response corresponding to the Content-Type of the request.

Genie.Renderer.registervars — Function

registervars(vars...) :: Nothing

Loads the rendering vars into the task's scope

Missing docstring.

Missing docstring for injectvars. Check Documenter's build log for details.

Genie.Renderer.view_file_info — Function

view_file_info(path::String, supported_extensions::Vector{String}) :: Tuple{String,String}

Extracts path and extension info about a file

Missing docstring.

Missing docstring for vars_signature. Check Documenter's build log for details.

Genie.Renderer.function_name — Function

function_name(file_path::String)

Generates function name for generated HTML+Julia views.

Genie.Renderer.m_name — Function

m_name(file_path::String)

Generates module name for generated HTML+Julia views.

Genie.Renderer.build_is_stale — Function

build_is_stale(file_path::String, build_path::String) :: Bool

Checks if the view template has been changed since the last time the template was compiled.

Genie.Renderer.build_module — Function

build_module(content::String, path::String, mod_name::String) :: String

Persists compiled Julia view data to file and returns the path

Genie.Renderer.preparebuilds — Function

preparebuilds() :: Bool

Sets up the build folder and the build module file for generating the compiled views.

Missing docstring.

Missing docstring for purgebuilds. Check Documenter's build log for details.

Missing docstring.

Missing docstring for changebuilds. Check Documenter's build log for details.

Missing docstring.

Missing docstring for vars. Check Documenter's build log for details.

Missing docstring.

Missing docstring for set_negotiated_content. Check Documenter's build log for details.

Genie.Renderer.negotiate_content — Function

negotiate_content(req::Request, res::Response, params::Params) :: Response

Computes the content-type of the Response, based on the information in the Request.