-

Genie

The highly productive Julia web framework

Documentation

Acknowledgements

Genie.Renderer.WebRenderable — Type.

mutable struct WebRenderable

Represents an object that can be rendered on the web

source

Genie.Renderer.tohtml — Function.

source

Genie.Renderer.html — Function.

source

Genie.Renderer.tojson — Function.

Invokes the JSON renderer of the underlying configured templating library.

source

Genie.Renderer.json — Function.

source

Genie.Renderer.redirect — Function.

Sets redirect headers and prepares the Response.

source

Genie.Renderer.hasrequested — Function.

hasrequested(content_type::Symbol) :: Bool

Checks wheter or not the requested content type matches content_type.

source

Genie.Renderer.respond — Function.

Constructs a Response corresponding to the content-type sof the request.

source